requirements.txt
python-telegram-bot
pdfplumber
pandas
openpyxl